Rename/tweak: skit_loaf_assign_slice -> skit_loaf_store_slice Similar functions were renamed too. skit_loaf_assign_cstr -> skit_loaf_store_cstr <array_loaf>_xfer_carr -> <array_loaf>_store_carr <array_loaf>_xfer_slice -> <array_loaf>_store_slice
The semantics of this function also changed slightly when a null
slice is assigned/stored into a non-null loaf:
(before) The loaf is freed.
(after)  Nothing happens and a null slice is returned.

THIS IS AN API BREAKING CHANGE.

Justification:
Bugs were introduced in calling code when numerous calls to
skit_loaf_assign_slice were assumed to shrink the loaf being
assigned to.  This was not the case, and it led to silent
corruption of string data.  This change should make such bugs
less likely by using a name change to emphasize that the function
is not an assignment, but merely a way to store some bytes into
an existing loaf.  It also makes the function more consistent:
it will now /never/ deallocate memory (ignoring technicalities
that happen inside the system's allocator when enlarging occurs
with realloc).

A later commit should create a new version of
skit_loaf_assign_slice that will actually turn the referenced loaf
into a carbon-copy of the slice given, (de)allocating as needed.