Bugfix for sRETURN/sSCOPE on OpenVMS.
Fix a bug that can cause it to appear as though values returned with
sRETURN(v) are being corrupted.  Actually, what is happening is the
corruption of random variables on the local stack.  Thus, this can
only happen if the value returned with sRETURN(...) is a variable in
the enclosing function's scope (or some expression of such variables).
(Other variables on the stack may be corrupted, but it doesn't matter
because the function is returning.)  This problem was present on
OpenVMS.  It could possibly happen on Linux, but that was never
observed.

This fix prevents the aforementioned corruption.  sRETURN should work
properly on OpenVMS now.