Allow the possibility of including a void* context when throwing exceptions.
This is necessary to replace some of the functionality that is lost when
exceptions aren't tied to heap-allocated class objects that can use
inheritence to create exception types with more detailed error
information.  In C Survival Kit, the void* context in sTHROWC can be
used to provide more detailed error information to whoever catches the
exception.

The ownership semantics of any memory that might need to
be allocated for the extra context information are currently vague and
undefined.  For now, it might be a good idea to just make it a pointer
to the object that threw the exception, if there is one.